feat(engine): Initial read isolation for transactions
Transactions now grab a handle to the current version when they are
opened. This is used to open readonly version of MVMap instances which
allows us to hide new stores and deletions from a transaction.

This change only implements this for get operations, stream and
indexes still do not support read transactions