feat(engine): Rework how indexes are kept up to date
Silo now uses an internal log to keep track of where a `QueryEngine` is
at. Engines are now required to keep track of their last hard commit, if
a crash occurs and an engine is not fully up to date Silo will replay
the log on the engine allowing it to become up to date without a full
rebuild.